unit test: formatNameTest was moved since the other class no longer exists
