+++ app/src/test/java/es/edufdezsoy/manga2kindle/service/ScanMangaTest.kt
@@ -175,4 +175,50 @@ class ScanMangaTest {
             assertEquals(results[i], res)
         }
     }
+
+    @Test
+    fun formatNameTest() {
+        println("\nTest: formatNameTest\n")
+
+        // Mock data
+        val names = arrayOf(
+            "Vol.TBD Chapter 104_ CSI Men",
+            "Chapter 105_ The Most Beautiful Girl on the Island",
+            "Service created_",
+            "Kyou no Yuiko-san",
+            "Karakai Jouzu no Takagi-san",
+            "Ookami_Shounen wa Kyou mo Uso o Kasaneru",
+            "Gokushufudou_ The Way of the House Husband",
+            "5Toubun no Hanayome",
+            "Kyou_no Yuiko-san_ testo",
+            "Capítulo 1.00 _ Una invocación real",
+            "Ch.56.0 _ Capítulo  56.00_  Sentimientos",
+            null
+        )
+
+        // Mock result
+        val results = arrayOf(
+            "Vol.TBD Chapter 104: CSI Men",
+            "Chapter 105: The Most Beautiful Girl on the Island",
+            "Service created_",
+            "Kyou no Yuiko-san",
+            "Karakai Jouzu no Takagi-san",
+            "Ookami_Shounen wa Kyou mo Uso o Kasaneru",
+            "Gokushufudou: The Way of the House Husband",
+            "5Toubun no Hanayome",
+            "Kyou_no Yuiko-san: testo",
+            "Capítulo 1.00 - Una invocación real",
+            "Ch.56.0 - Capítulo 56.00: Sentimientos",
+            ""
+        )
+
+        for (i in names.indices) {
+            val res = scanManga.formatName(names[i])
+
+            println("Asserting: " + results[i] + " (Input: " + names[i] + ")")
+            println("with:      " + res)
+
+            assertEquals(results[i], res)
+        }
+    }
 }
